requirements.txt
# ═══════════════════════════════════════════════════════════════════
#  AeroMind Sentinel — Python Dependencies
#  Install: pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════════════

# ─── Core ─────────────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# ─── Deep Learning ────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ─── Detection ────────────────────────────────────────────────────
ultralytics>=8.0.0
sahi>=0.11.0

# ─── Tracking ─────────────────────────────────────────────────────
filterpy>=1.4.5
lap>=0.4.0

# ─── Reinforcement Learning ───────────────────────────────────────
stable-baselines3>=2.1.0
gymnasium>=0.29.0

# ─── AirSim RPC ──────────────────────────────────────────────────
msgpack>=1.0.5

# ─── Metrics & Evaluation ────────────────────────────────────────
motmetrics>=1.4.0

# ─── Visualization ───────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# ─── Utilities ───────────────────────────────────────────────────
pyyaml>=6.0
loguru>=0.7.0
rich>=13.0.0
omegaconf>=2.3.0

# ─── Dashboard ───────────────────────────────────────────────────
flask>=3.0.0

# ─── Testing ─────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
